feat: add authentication buttons to unauthorized modal
