+++ public/locales/en/translation.json
@@ -23,8 +23,8 @@
       "errorModalTitle": "Processing failed",
       "errorModalText": "The network is unstable and it was not possible to process your donation. Try again shortly.",
       "errorModalButtonText": "Ok",
-      "unauthorizedModalTitle": "Validate your e-mail",
-      "unauthorizedModalText": "Please, validate your e-mail to ensure your donation arrives safely.",
+      "unauthorizedModalTitle": "Sign in to continue donating",
+      "unauthorizedModalText": "Please, sign in to your account created with {{email}} to continue your donation.",
       "unauthorizedModalButtonText": "validate my account",
       "receiveTicketAnimationModalTitle": "Receiving Ticket...",
       "receiveTicketAnimationModalOrigin": "Community",
